Use glm
Using my own implementation of Vec3 was fun for learning, but using a
standardized library for algebra is safer and I don't have to test it.

I might also have to use other objects such as matrices in the future, which
are also provided by the library.

It was surprisingly easy to substitute my Vec3 with glm::DVec3.